move menu component to common folder
