+++ bigbluebutton-html5/imports/ui/components/common/menu/component.jsx
